# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BrewDesk Café Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff members
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a new staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
  /staff/{id}:
    get:
      operationId: getStaff
      tags: [staff]
      summary: Get a staff member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Tables ────────────────────────────────────────────────────────────
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all rooms and halls
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new room or hall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get a room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    patch:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Menu Categories ───────────────────────────────────────────────────
...
  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Get revenue and expense summary by period
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Report summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
  /reports/transactions:
    get:
      operationId: getTransactionsReport
      tags: [reports]
      summary: Full transactions log with line items and breakdowns for a period
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Transactions report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionsReport"
...
  /reports/cancellations:
    get:
      operationId: getCancellationsReport
      tags: [reports]
      summary: Cancelled/deleted orders and cancelled tables/rooms (manager only)
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Cancellations report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancellationsReport"
        "403":
          description: Manager access required

  # ── Dashboard ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Staff:
      type: object
      required: [id, name, username, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, manager, cashier, waiter, kitchen]
        isActive:
          type: boolean
        phone:
          type: ["string", "null"]
        permissions:
          type: ["object", "null"]
        createdAt:
          type: string
    StaffInput:
      type: object
      required: [name, username, password, role]
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, manager, cashier, waiter, kitchen]
        phone:
          type: string
        permissions:
          type: object
    StaffUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, cashier, waiter, kitchen]
        phone:
          type: string
        isActive:
          type: boolean
        permissions:
          type: object
        password:
          type: string

    # ── Tables ──────────────────────────────────────────────────────────
...
    Room:
      type: object
      required: [id, name, capacity, type, status, hourlyRate]
      properties:
        id:
          type: integer
        name:
          type: string
        capacity:
          type: integer
        type:
          type: string
          enum: [meeting, lecture, training]
        status:
          type: string
          enum: [available, reserved, occupied]
        hourlyRate:
          type: number
          minimum: 0
        hasProjector:
          type: boolean
        hasMicrophone:
          type: boolean
        hasWhiteboard:
          type: boolean
        qrCode:
          type: ["string", "null"]
    RoomInput:
      type: object
      required: [name, capacity, type, hourlyRate]
      properties:
        name:
          type: string
        capacity:
          type: integer
        type:
          type: string
          enum: [meeting, lecture, training]
        hourlyRate:
          type: number
        hasProjector:
          type: boolean
        hasMicrophone:
          type: boolean
        hasWhiteboard:
          type: boolean
    RoomUpdate:
      type: object
      properties:
        name:
          type: string
        capacity:
          type: integer
        type:
          type: string
          enum: [meeting, lecture, training]
        status:
          type: string
          enum: [available, reserved, occupied]
        hourlyRate:
          type: number
          minimum: 0
        hasProjector:
          type: boolean
        hasMicrophone:
          type: boolean
        hasWhiteboard:
          type: boolean

    # ── Menu ────────────────────────────────────────────────────────────
...
    InvoiceItem:
      type: object
      required: [description, quantity, unitPrice, subtotal]
      properties:
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
...
    ReportSummary:
      type: object
      required: [period, totalRevenue, totalExpenses, netProfit, invoiceCount, guestCount, avgOrderValue, revenueByDay]
      properties:
        period:
          type: string
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        invoiceCount:
          type: integer
        guestCount:
          type: integer
        avgOrderValue:
          type: number
        revenueByDay:
          type: array
          items:
            type: object
            required: [date, revenue, expenses]
            properties:
              date:
                type: string
              revenue:
                type: number
              expenses:
                type: number
    TransactionsReport:
      type: object
      required: [period, start, end, invoiceCount, totalRevenue, timeRevenue, productRevenue, adjustments, totalDiscounts, topItems, details]
      properties:
        period:
          type: string
        start:
          type: string
        end:
          type: string
        invoiceCount:
          type: integer
        totalRevenue:
          type: number
        timeRevenue:
          type: number
        productRevenue:
          type: number
        adjustments:
          type: number
        totalDiscounts:
          type: number
        topItems:
          type: array
          items:
            type: object
            required: [name, quantity, total]
            properties:
              name:
                type: string
              quantity:
                type: number
              total:
                type: number
        details:
          type: array
          items:
            $ref: "#/components/schemas/TransactionDetail"
    TransactionDetail:
      type: object
      required: [id, createdAt, subtotal, total, items]
      properties:
        id:
          type: integer
        createdAt:
          type: string
        sessionId:
          type: ["integer", "null"]
        resourceName:
          type: ["string", "null"]
        sessionType:
          type: ["string", "null"]
        guestCount:
          type: ["integer", "null"]
        sessionStart:
          type: ["string", "null"]
        sessionEnd:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        staffName:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        subtotal:
          type: number
        discountAmount:
          type: ["number", "null"]
        total:
          type: number
        amountPaid:
          type: ["number", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
...
    CancellationTally:
      type: object
      required: [count, total]
      properties:
        count:
          type: integer
        total:
          type: number
    CancellationDetail:
      type: object
      required: [id, kind, action, label, amount, createdAt]
      properties:
        id:
          type: integer
        kind:
          type: string
          enum: [order, order_item, session, guest]
        action:
          type: string
          enum: [cancelled, deleted, removed]
        resourceType:
          type: ["string", "null"]
        label:
          type: string
        amount:
          type: number
        reason:
          type: ["string", "null"]
        staffName:
          type: ["string", "null"]
        createdAt:
          type: string
    CancellationsReport:
      type: object
      required: [period, orders, items, tables, rooms, guests, grandTotal, details]
      properties:
        period:
          type: string
        orders:
          $ref: "#/components/schemas/CancellationTally"
        items:
          $ref: "#/components/schemas/CancellationTally"
        tables:
          $ref: "#/components/schemas/CancellationTally"
        rooms:
          $ref: "#/components/schemas/CancellationTally"
        guests:
          $ref: "#/components/schemas/CancellationTally"
        grandTotal:
          type: number
        details:
          type: array
          items:
            $ref: "#/components/schemas/CancellationDetail"